+++ test/unit/fuzz-tools.test.js
@@ -129,6 +129,7 @@ import * as renalV128 from '../../lib/renal-v128.js';
 import * as acidbaseV129 from '../../lib/acidbase-v129.js';
 import * as uroV130 from '../../lib/uro-v130.js';
 import * as uroV131 from '../../lib/uro-v131.js';
+import * as hemeV132 from '../../lib/heme-v132.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -203,6 +204,7 @@ const MODULES = {
   'acidbase-v129.js': acidbaseV129,
   'uro-v130.js': uroV130,
   'uro-v131.js': uroV131,
+  'heme-v132.js': hemeV132,
   'ems-v149.js': emsV149,
 };